requirements.txt
flask==1.1.1
flask_pymongo==2.3.0
python-dotenv==0.12.0
simplejson==3.17.0
